requirements.txt.txt
flask
python-telegram-bot
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn